requirements_enhanced.txt
# Core LangChain and LangGraph dependencies (using compatible versions)
langchain==0.3.27
langchain-core==0.3.72
langchain-community==0.3.13
langchain-openai==0.3.28
langgraph==0.6.1
langgraph-checkpoint-sqlite==2.0.11

# Streamlit for web interface
streamlit==1.47.1

# Database and storage (sqlite3 is built into Python)
aiosqlite==0.21.0

# Environment management
python-dotenv==1.1.1

# HTTP requests and APIs
requests==2.32.4
httpx==0.28.1

# Data processing and utilities
pandas==2.3.1
numpy==2.3.2
pydantic==2.11.7

# Social media APIs (for future integration)
tweepy==4.14.0
python-linkedin-v2==1.1.5

# YouTube and video processing
youtube-transcript-api==0.6.2
pytube==15.0.0

# Additional AI/ML libraries
openai==1.98.0
anthropic==0.40.0

# Async support
aiofiles==24.1.0

# Logging and monitoring
loguru==0.7.2

# Development and testing
pytest==8.3.4
black==24.10.0
flake8==7.1.1

# Web scraping (for enhanced search)
beautifulsoup4==4.12.3
selenium==4.27.1

# Image processing (for future multimodal features)
Pillow==11.3.0
opencv-python==4.10.0.84

# Document processing
pypdf2==3.0.1
python-docx==1.1.2

# Scheduling and task management
celery==5.4.0
redis==5.2.1

# API framework (for future API endpoints)
fastapi==0.115.6
uvicorn==0.32.1

# Monitoring and observability
prometheus-client==0.21.1

# Additional utilities
click==8.2.1
colorama==0.4.6
tqdm==4.67.1
jsonschema==4.25.0